pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lean-ai-serve"
version = "0.1.0"
description = "Secure vLLM inference server with model management, fine-tuning, and HIPAA-grade audit logging"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Lean AI" },
]
keywords = ["llm", "vllm", "inference", "model-management", "fine-tuning", "enterprise"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Web framework
    "fastapi>=0.115.0,<1.0",
    "uvicorn[standard]>=0.32.0,<1.0",

    # Async HTTP client (reverse proxy to vLLM)
    "httpx>=0.27.0,<1.0",

    # Database
    "aiosqlite>=0.20.0,<1.0",

    # Configuration
    "pyyaml>=6.0",
    "pydantic>=2.9.0,<3.0",
    "pydantic-settings>=2.6.0,<3.0",

    # HuggingFace model management
    "huggingface-hub>=0.26.0,<1.0",

    # Process management
    "psutil>=6.0.0",

    # Security
    "cryptography>=43.0.0",
    "pyjwt[crypto]>=2.9.0,<3.0",
    "bcrypt>=4.2.0,<5.0",

    # Structured logging
    "structlog>=24.4.0",

    # CLI
    "typer>=0.12.0,<1.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
ldap = [
    "ldap3>=2.9.1,<3.0",
]
gpu = [
    "nvidia-ml-py>=12.560.0",
]
compression = [
    "llmlingua>=0.2.0",
]
training = [
    # LLaMA-Factory is installed separately (complex deps)
    # This group provides dataset handling utilities
    "pandas>=2.2.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "httpx",  # for TestClient
]

[project.scripts]
lean-ai-serve = "lean_ai_serve.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/lean_ai_serve"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]
ignore = ["B008"]  # Depends() in FastAPI defaults is standard pattern

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
